% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import.R
\name{n_datasets}
\alias{n_datasets}
\title{Count number of datasets in a FCS file}
\usage{
n_datasets(filename)
}
\arguments{
\item{filename}{FCS filename,}
}
\description{
Count number of datasets in a FCS file
}
